fix /= operator (2)

+++ int128v1.hpp
@@ -226,7 +226,7 @@ uint128_t &operator-=(uint128_t &a,uint128_t b){
 uint128_t &operator*=(uint128_t &a,uint128_t b){
     return a = a * b;
 }
-uint128_t &operator/=(uint128_t &a,unsigned int b){
+uint128_t &operator/=(uint128_t &a,unsigned long long b){
     return a = a / b;
 }
 uint128_t &operator%=(uint128_t &a,unsigned long long b){
